Refactored the appropriate table operations for DA to the right class. Cleaned up the DataGetter class
