+++ DataAccess/FutureGames/IFutureGamesDA.cs
@@ -5,5 +5,6 @@ namespace DataAccess.FutureGames
     public interface IFutureGamesDA
     {
         void AddFutureGames(List<FutureGame> games);
+        List<FutureGame> GetFutureGames();
     }
 }
